apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv
openai
alembic
python-multipart
PyMuPDF
python-docx
python-pptx